Module Yocaml_jingoo

Plugin for describing templates using Jingoo, a template engine inspired by Jinja2 (from the Python world).

Injection

Description of a pipeline for reading a template and injecting content.

module Pipeline : sig ... end

Describes the arrowized interface for reading a file as a template and injecting content and variables.

Data template

Describes the Yocaml_jingoo module as a template engine. Allows the latter to be passed to any function requiring it.

type t = Jingoo.Jg_types.tvalue

The type that describes the template language.

val from : Yocaml.Data.t -> t

from data Transforms a normalized data representation (data) into an associative list of data that can be injected into a template.

val render : ?strict:bool -> (string * t) list -> string -> string

render ?strict parameters content injects parameters data into content and returns the result of the applied content. To inject metadata into a template.